refactored the View controller and moved each view in a separate file
